4 - standardize calculation for minimum transaction amount (#48)
* standardize calculation for minimum transaction amount

Hyperdrive returns the minimum transaction amount in base and rounds up
when handling conversions.

Previously, it was not being converted in some places and values were
rounded down.

This pr unifies the minimum transaction amount calculation and adds a
buffer to overestimate it.

* use strategy's minimum transaction amount instead of constant

* responding to feedback from @mrtoph

Remove additional conversion for wrapped tokens since the wrapped token
denominations are the same as hyperdrive's vault shares token.