5 - closePositions respects minimumTransactionAmount (depends on 4) (#49)
* standardize calculation for minimum transaction amount

Hyperdrive returns the minimum transaction amount in base and rounds up
when handling conversions.

Previously, it was not being converted in some places and values were
rounded down.

This pr unifies the minimum transaction amount calculation and adds a
buffer to overestimate it.

* use strategy's minimum transaction amount instead of constant

* responding to feedback from @mrtoph

Remove additional conversion for wrapped tokens since the wrapped token
denominations are the same as hyperdrive's vault shares token.

* 5 - closePositions respects minimumTransactionAmount

Currently, `closePositions(..)` does not check whether a partial closure
will result in a position value less than the minimum transaction
amount.

This PR adds the necessary logic for this check as well as a test to
ensure it is behaving as expected.

[issue](https://cantina.xyz/code/4f25dfd5-d3e6-4e7a-9481-d7306b795f2b/findings/5)

* commenting

* add test with asBase=false

* addressing feedback from @mrtoph

No conversions are needed for min tx amount when closing longs since
it's denominated in bonds not execution tokens

* pin foundry version to a stable version

* pin foundry version to a stable version

---------

Co-authored-by: Chad0 <Sean329@users.noreply.github.com>
Co-authored-by: Xiangyu Xu <xiangyuxu329@gmail.com>